rattler_repodata_gateway

Module fetch

This module provides functionality to download and cache repodata.json from a remote location.

Modules

no_cache

Provides methods to download repodata from a given channel URL but does not perform any form of caching.

Structs

CachedRepoData

The result of fetch_repo_data.

FetchRepoDataOptions

Additional knobs that allow you to tweak the behavior of fetch_repo_data.

VariantAvailability

Describes the availability of certain repodata.json.

Enums

CacheAction

Defines how to use the repodata cache.

CacheResult

Indicates whether or not the repodata.json cache was up-to-date or not.

FetchRepoDataError

RepoDataNotFoundError

RepoData could not be found for given channel and platform

Variant

Defines which type of repodata.json file to download. Usually you want to use the Variant::AfterPatches variant because that reflects the repodata with any patches applied.

Functions

check_variant_availability

Determine the availability of repodata.json variants (like a .zst or .bz2) by checking a cache or the internet.

fetch_repo_data

Fetch the repodata.json file for the given subdirectory. The result is cached on disk using the HTTP cache headers returned from the server.