rattler_repodata_gateway::fetch

Function fetch_repo_data

pub async fn fetch_repo_data(
    subdir_url: Url,
    client: LazyClient,
    cache_path: PathBuf,
    options: FetchRepoDataOptions,
    reporter: Option<Arc<dyn Reporter>>,
) -> Result<CachedRepoData, FetchRepoDataError>

Fetch the repodata.json file for the given subdirectory. The result is cached on disk using the HTTP cache headers returned from the server.

The successful result of this function also returns a lockfile which ensures that both the state and the repodata that is pointed to remain in sync. However, not releasing the lockfile (by dropping it) could block other threads and processes, it is therefore advisable to release it as quickly as possible.

This method implements several different methods to download the repodata.json file from the remote:

If a repodata.json.zst file is available in the same directory that file is downloaded and decompressed.

If a repodata.json.bz2 file is available in the same directory that file is downloaded and decompressed.

Otherwise the regular repodata.json file is downloaded.

The checks to see if a .zst and/or .bz2 file exist are performed by doing a HEAD request to the respective URLs. The result of these are cached.