rattler_repodata_gateway::sparse

Function load_repo_data_recursively

pub async fn load_repo_data_recursively(
    repo_data_paths: impl IntoIterator<Item = (Channel, impl Into<String>, impl AsRef<Path>)>,
    package_names: impl IntoIterator<Item = PackageName>,
    patch_function: Option<fn(&mut PackageRecord)>,
    variant_consolidation: PackageFormatSelection,
) -> Result<Vec<Vec<RepoDataRecord>>, Error>

A helper function that immediately loads the records for the given packages (and their dependencies). Records for the specified packages are loaded from the repodata files. The patch_record_fn is applied to each record after it has been parsed and can mutate the record after it has been loaded.